+++ __tests__/test-GitObjectManager.js
@@ -2,7 +2,7 @@
 const { makeFixture } = require('./__helpers__/FixtureFS.js')
 const snapshots = require('./__snapshots__/test-GitObjectManager.js.snap')
 const registerSnapshots = require('./__helpers__/jasmine-snapshots')
-const GitObjectManager = require('isomorphic-git/internal-apis')
+const { readObject } = require('isomorphic-git/internal-apis')
 
 describe('GitObjectManager', () => {
   beforeAll(() => {
@@ -15,7 +15,7 @@ describe('GitObjectManager', () => {
     // Test
     let error = null
     try {
-      await GitObjectManager.readObject({
+      await readObject({
         fs,
         gitdir,
         oid: 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
