+++ src/commands/expandOid.js
@@ -1,6 +1,6 @@
 import path from 'path'
 
-import { expandOid as _expandOid } from '../managers/GitObjectManager.js'
+import { expandOid as _expandOid } from '../utils/expandOid.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { cores } from '../utils/plugins.js'
