+++ src/commands/readObject.js
@@ -1,11 +1,11 @@
 import path from 'path'
 
-import { readObject as _readObject } from '../managers/GitObjectManager.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { GitAnnotatedTag } from '../models/GitAnnotatedTag.js'
 import { GitCommit } from '../models/GitCommit.js'
 import { E, GitError } from '../models/GitError.js'
 import { GitTree } from '../models/GitTree.js'
+import { readObject as _readObject } from '../utils/readObject.js'
 import { resolveTree } from '../utils/resolveTree.js'
 import { cores } from '../utils/plugins.js'
