+++ src/commands/resetIndex.js
@@ -1,12 +1,12 @@
 import path from 'path'
 
 import { GitIndexManager } from '../managers/GitIndexManager.js'
-import { hash } from '../managers/GitObjectManager.js'
 import { GitRefManager } from '../managers/GitRefManager.js'
 import { FileSystem } from '../models/FileSystem.js'
+import { hashObject } from '../utils/hashObject.js'
 import { cores } from '../utils/plugins.js'
 
-import { readObject } from './readObject'
+import { readObject } from './readObject.js'
 
 /**
  * Reset a file in the git index (aka staging area)
@@ -49,7 +49,7 @@ export async function resetIndex ({
     const object = await fs.read(path.join(dir, filepath))
     if (object) {
       // ... and has the same hash as the desired state...
-      workdirOid = await hash({
+      workdirOid = await hashObject({
         gitdir,
         type: 'blob',
         object
