+++ src/commands/sign.js
@@ -1,10 +1,11 @@
 import path from 'path'
 
-import { readObject, writeObject } from '../managers/GitObjectManager.js'
 import { GitRefManager } from '../managers/GitRefManager.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { E, GitError } from '../models/GitError.js'
 import { SignedGitCommit } from '../models/SignedGitCommit.js'
+import { readObject } from '../utils/readObject.js'
+import { writeObject } from '../utils/writeObject.js'
 import { cores } from '../utils/plugins.js'
 
 /**
