+++ src/internal-apis.js
@@ -5,7 +5,6 @@ export * from './commands/pack'
 export * from './managers/GitConfigManager'
 export * from './managers/GitIgnoreManager'
 export * from './managers/GitIndexManager'
-export * from './managers/GitObjectManager'
 export * from './managers/GitRefManager'
 export * from './managers/GitRemoteConnection'
 export * from './managers/GitRemoteHTTP'
@@ -38,8 +37,10 @@ export * from './utils/padHex'
 export * from './utils/path'
 export * from './utils/pkg'
 export * from './utils/plugins'
+export * from './utils/readObject'
 export * from './utils/resolveTree'
 export * from './utils/shasum'
 export * from './utils/sleep'
 export * from './utils/symbols'
 export * from './utils/t'
+export * from './utils/writeObject'
