+++ src/utils/readObject.js
@@ -1,7 +1,7 @@
 import pako from 'pako'
 
-import * as GitObjectStoreLoose from '../managers/GitObjectStoreLoose.js'
-import * as GitObjectStorePacked from '../managers/GitObjectStorePacked.js'
+import { read as readLoose } from '../managers/GitObjectStoreLoose.js'
+import { read as readPacked } from '../managers/GitObjectStorePacked.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { E, GitError } from '../models/GitError.js'
 import { GitObject } from '../models/GitObject.js'
@@ -14,10 +14,10 @@ export async function readObject ({ fs: _fs, gitdir, oid, format = 'content' }) 
   const getExternalRefDelta = oid => readObject({ fs: _fs, gitdir, oid })
 
   // Look for it in the loose object directory.
-  let result = await GitObjectStoreLoose.read({ fs, gitdir, oid })
+  let result = await readLoose({ fs, gitdir, oid })
   // Check to see if it's in a packfile.
   if (!result) {
-    result = await GitObjectStorePacked.read({ fs, gitdir, oid, getExternalRefDelta })
+    result = await readPacked({ fs, gitdir, oid, getExternalRefDelta })
   }
   // Finally
   if (!result) {
@@ -54,43 +54,4 @@ export async function readObject ({ fs: _fs, gitdir, oid, format = 'content' }) 
       })
   }
   /* eslint-enable no-fallthrough */
-}
-
-export async function hash ({ gitdir, type, object }) {
-  return shasum(GitObject.wrap({ type, object }))
-}
-
-export async function expandOid ({ fs: _fs, gitdir, oid: short }) {
-  const fs = new FileSystem(_fs)
-  // Curry the current read method so that the packfile un-deltification
-  // process can acquire external ref-deltas.
-  const getExternalRefDelta = oid => readObject({ fs: _fs, gitdir, oid })
-
-  const results1 = await GitObjectStoreLoose.expandOid({ fs, gitdir, oid: short })
-  const results2 = await GitObjectStorePacked.expandOid({ fs, gitdir, oid: short, getExternalRefDelta })
-  const results = results1.concat(results2)
-
-  if (results.length === 1) {
-    return results[0]
-  }
-  if (results.length > 1) {
-    throw new GitError(E.AmbiguousShortOid, {
-      short,
-      matches: results.join(', ')
-    })
-  }
-  throw new GitError(E.ShortOidNotFound, { short })
-}
-
-export async function writeObject ({ fs: _fs, gitdir, type, object, format = 'content', oid }) {
-  const fs = new FileSystem(_fs)
-  if (format !== 'deflated') {
-    if (format !== 'wrapped') {
-      object = GitObject.wrap({ type, object })
-    }
-    oid = shasum(object)
-    object = Buffer.from(pako.deflate(object))
-  }
-  await GitObjectStoreLoose.write({ fs, gitdir, object, format: 'deflated', oid })
-  return oid
 }
