+++ src/utils/resolveTree.js
@@ -1,4 +1,4 @@
-import { readObject } from '../managers/GitObjectManager.js'
+import { readObject } from '../utils/readObject.js'
 import { GitAnnotatedTag } from '../models/GitAnnotatedTag.js'
 import { GitCommit } from '../models/GitCommit.js'
 import { E, GitError } from '../models/GitError.js'
